+++ loader/include/Geode/loader/IPC.hpp
@@ -4,9 +4,9 @@
 #include "Loader.hpp"
 #include <matjson.hpp>
 
-namespace geode {
+namespace geode::ipc {
     #ifdef GEODE_IS_WINDOWS
-    constexpr char const* IPC_PIPE_NAME = "\\\\.\\pipe\\GeodeIPCPipe";
+    constexpr char const* IPC_PIPE_NAME = R"(\\.\pipe\GeodeIPCPipe)";
     #endif
 
     #ifdef GEODE_IS_MACOS
@@ -15,20 +15,20 @@ namespace geode {
 
     class IPCFilter;
 
-    // IPC (Inter-Process Communication) provides a way for Geode mods to talk 
-    // to other programs on the user's computer. If you have, for example, a 
-    // debugger, or an external modding UI, that application can open up the 
-    // platform-specific pipe and start sending messages to mods. Mods can 
-    // listen for messages using the listenForIPC function, and reply to 
-    // messages the get by using the reply method on the event provided. For 
-    // example, an external application can query what mods are loaded in Geode 
+    // IPC (Inter-Process Communication) provides a way for Geode mods to talk
+    // to other programs on the user's computer. If you have, for example, a
+    // debugger, or an external modding UI, that application can open up the
+    // platform-specific pipe and start sending messages to mods. Mods can
+    // listen for messages using the listenForIPC function, and reply to
+    // messages the get by using the reply method on the event provided. For
+    // example, an external application can query what mods are loaded in Geode
     // by sending the `list-mods` message to `geode.loader`.
 
     class GEODE_DLL IPCEvent : public Event {
     protected:
         void* m_rawPipeHandle;
         bool m_replied = false;
-    
+
     public:
         std::string targetModID;
         std::string messageID;
@@ -57,12 +57,12 @@ namespace geode {
 
     public:
         ListenerResult handle(utils::MiniFunction<Callback> fn, IPCEvent* event);
-		IPCFilter(
+        IPCFilter(
             std::string const& modID,
             std::string const& messageID
         );
         IPCFilter(IPCFilter const&) = default;
     };
 
-    std::monostate listenForIPC(std::string const& messageID, matjson::Value(*callback)(IPCEvent*));
+    std::monostate listen(std::string const& messageID, matjson::Value(*callback)(IPCEvent*));
 }
