+++ loader/src/platform/android/IPC.cpp
@@ -1,0 +1,7 @@
+﻿#include <loader/IPC.hpp>
+
+using namespace geode::prelude;
+
+void ipc::setup() {
+    log::debug("IPC is not supported on this platform!");
+}
